// Frontend/app/api/generated/model/estadoHistoricoResponseDTOEstado.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type EstadoHistoricoResponseDTOEstado = typeof EstadoHistoricoResponseDTOEstado[keyof typeof EstadoHistoricoResponseDTOEstado];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EstadoHistoricoResponseDTOEstado = {
  INCOMPLETO_POR_ADMINISTRACION: 'INCOMPLETO_POR_ADMINISTRACION',
  RECHAZADO_A_ADMINISTRACION: 'RECHAZADO_A_ADMINISTRACION',
  COMPLETO_POR_ADMINISTRACION: 'COMPLETO_POR_ADMINISTRACION',
  INCOMPLETO_POR_PROFESOR: 'INCOMPLETO_POR_PROFESOR',
  RECHAZADO_A_PROFESOR: 'RECHAZADO_A_PROFESOR',
  COMPLETO_POR_PROFESOR: 'COMPLETO_POR_PROFESOR',
  APROBADO_POR_COMISION: 'APROBADO_POR_COMISION',
  APROBADO_POR_SECRETARIA: 'APROBADO_POR_SECRETARIA',
} as const;